DOC: Add axes to the viz_bundles example
This commit adds an `axes` actor to one example from the viz_bundles.py
example. This introduces the user to the following VTK concepts:

- the `axes` actor
- the ability to query the position and range of existing actors
- the ability to set the position and scale of existing actors

Axes are commonly used to communicate to the viewer the orientation of
the brain volume or slice. It took me a while to learn how to do this
for my figures and I thought it would be helpful to the DIPY tyro if the
proper incantations were documented here.